Rename grammar_ast.rs to ast.rs.
I was continually confused between grammar.rs and grammar_ast.rs. This rename
might help, or maybe I'm beyond help.
